Common is more of a demo resources packagse, so rename!
