+++ demo-resources/src/main/res/layout/activity_demo.xml
