+++ demo-resources/src/main/res/values/colors.xml
