+++ demo-resources/src/main/res/values/strings.xml
