+++ demo-resources/src/main/res/values/styles.xml
